Refactored new binary search code
