+++ spotlight/commands/search_cache.py
@@ -3,7 +3,7 @@ from spotlight.suggestions.playable.album import AlbumSuggestion
 from spotlight.suggestions.playable.artist import ArtistSuggestion
 from spotlight.suggestions.playable.playlist import PlaylistSuggestion
 from spotlight.suggestions.playable.song import SongSuggestion, QueueSuggestion
-from spotlight.suggestions.templates import FillSuggestion, PassiveSuggestion
+from spotlight.suggestions.templates import FillSuggestion
 from spotlight.commands.command import Command
 
 
@@ -31,32 +31,59 @@ class SearchCacheCommand(Command):
             return [FillSuggestion(self.title, self.description, self.prefix[:-1], self.prefix)]
 
         online_item = FillSuggestion(f"Search Online for '{parameter}'", "Search Online", "search", f"🔎{self.type_} {parameter}")
-        item_list, title, image, item, cache, description = [online_item], \
+        result_list, title, image, suggestion, cache, description = [online_item], \
                                                             "name", "image", None, None, "description"
 
         if self.type_ == "song" or self.type_ == "queue":
             description = "artist"
             cache = CacheHolder.song_cache
-            item = SongSuggestion if self.type_ == "song" else QueueSuggestion
+            suggestion = SongSuggestion if self.type_ == "song" else QueueSuggestion
         if self.type_ == "playlist":
             description = "owner"
             cache = CacheHolder.playlist_cache
-            item = PlaylistSuggestion
+            suggestion = PlaylistSuggestion
         if self.type_ == "album":
             cache = CacheHolder.album_cache
             description = "artist"
-            item = AlbumSuggestion
+            suggestion = AlbumSuggestion
         if self.type_ == "artist":
             cache = CacheHolder.artist_cache
             description = "genre"
-            item = ArtistSuggestion
+            suggestion = ArtistSuggestion
 
-        for key, values in cache[f'{self.type_ if self.type_ != "queue" else "song"}s'].items():
-            name = values[title]
-            if len(item_list) == 6:
+        song_list = list(cache[f'{self.type_ if self.type_ != "queue" else "song"}s'].items())
+        song_list = sorted(song_list, key=self.clean_name)
+
+        while len(result_list) < 6:
+            result = self.binary_search(song_list, parameter)
+            if result:
+                key, values = song_list[result]
+                new_suggestion = suggestion(values[title], values[description], values[image], key)
+                result_list.append(new_suggestion)
+                song_list.pop(result)
+            else:
                 break
-            if len(name) >= len(parameter) and name[:len(parameter)].lower() == parameter:
-                new_suggestion = item(name, values[description], values[image], key)
-                item_list.append(new_suggestion)
-                # TODO: Add duplicate removal system and change search from linear to binary
-        return item_list
+        return result_list
+
+    def clean_name(self, x):
+        name = x.lower() if isinstance(x,str) else x[1]["name"].lower()
+        for i in ["@", "'", '"', "¡", "!", "#", "$", "%", ".", ","]:
+            name = name.replace(i, "")
+        return name
+
+    def binary_search(self, L, target):
+        start = 0
+        end = len(L) - 1
+
+        while start <= end:
+            middle = (start + end)// 2
+            search = self.clean_name(L[middle])
+            target = self.clean_name(target)
+
+            if len(search)>=len(target) and search[:len(target)].lower() == target.lower():
+                return middle
+
+            if search.lower() > target.lower():
+                end = middle - 1
+            elif search.lower() < target.lower():
+                start = middle + 1
